test: static 메소드 위치 변경
